# web/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Driver API",
        "description": "For auth you may use token=`100-token` (without quotes)",
        "version": "0.1"
    },
    "paths": {
...
        "/model": {
            "get": {
                "tags": [
                    "Models of bus"
                ],
                "summary": "List of bus models",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bus models",
                        "headers": {
                            "X-Pagination-Current-Page": {
                                "description": "Current page",
                                "schema": {
                                    "type": "int"
                                }
                            },
                            "X-Pagination-Page-Count": {
                                "description": "Count of pages",
                                "schema": {
                                    "type": "int"
                                }
                            },
                            "X-Pagination-Per-Page": {
                                "description": "Count elements of page",
                                "schema": {
                                    "type": "int"
                                }
                            },
                            "X-Pagination-Total-Count": {
                                "description": "Total count of elements",
                                "schema": {
                                    "type": "int"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/model"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Models of bus"
                ],
                "summary": "Update one bus model",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/model"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated bus model info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Models of bus"
                ],
                "summary": "Create one bus model",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/modelCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created bus model info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/model/{id}": {
            "get": {
                "tags": [
                    "Models of bus"
                ],
                "summary": "One bus model info",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of bus model to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One bus model info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "model": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "speed": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Modelname",
                    "speed": 60
                }
            },
            "modelCreate": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "speed": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Modelname",
                    "speed": 60
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            },
            "tokenAuth": {
                "type": "apiKey",
                "name": "token",
                "in": "query",
                "scheme": "token"
            }
        }
    }
}